+++ src/sdk/CMakeLists.txt
@@ -203,44 +203,34 @@ target_include_directories (
 
 # To get the final SDK, we need some libraries
 # Below, we will find and integrate these required libraries
+# LIBRARIES_PATH save all required libraries path
 
-# We need hybridse_flags, but it hasn't been mentioned and needs to be added manually
-list(APPEND required_libs ${BIN_LIBS} ${SDK_LIBS} ${LINK_LIBS} hybridse_flags)
+# Get path by $<TARGET_FILE:${X}> for some libraries built in OpenMLDB
+set(SELF_LIBS ${BUILTIN_LIBS} openmldb_sdk openmldb_flags hybridse_sdk hybridse_core hybridse_flags)
+foreach(X  IN LISTS SELF_LIBS)
+    list(APPEND LIBRARIES_PATH $<TARGET_FILE:${X}>)
+endforeach()
 
-# In required_libs, some libraries do not need or require different path lookup methods
-# ${ABSL_LIBS} ${LLVM_LIBS} : we need to recursively look up all the libraries they depend on
-# ${file_based_test_driver_LIBS} : we don't need the libraries, so we remove them
-# unwind : in ${LINK_LIBS}, we can get it by ${UNWIND_LIBRARY} which in ${BRPC_LIBS}
-# common tcmalloc_minimal : in ${BIN_LIBS}, we don't need them
-# zookeeper_mt : we get it by ${zookeeper_mt_LIBRARY}
-# zetasql : we get it by ${zetasql_LIBRARY}
-# re2 : we get it by ${re2_LIBRARY}
-# date_proto timeofday_proto latlng_proto : we don't need them
-# -lpthread : in zookeeper_mt
-# dl :we get it by ${dl_LIBRARY}
-# pthread : we get it by ${pthread_LIBRARY}
-# rt : we get it by ${rt_LIBRARY}
-list(REMOVE_ITEM required_libs ${ABSL_LIBS} ${LLVM_LIBS} ${file_based_test_driver_LIBS} unwind    common tcmalloc_minimal zookeeper_mt    zetasql re2 date_proto timeofday_proto latlng_proto    -lpthread dl pthread rt)
-
-# Process libraries in required_libs :
-# 1. with path itself
-# 2. get path by $<TARGET_FILE:${X}>
-# libraries_path sava paths of  the all libraries
-foreach(X  IN LISTS required_libs)
+list(APPEND MERGED_LIBS ${BIN_LIBS} ${SDK_LIBS} ${LINK_LIBS})
+# Some libraries have their own paths
+foreach(X  IN LISTS MERGED_LIBS)
     if(EXISTS ${X})
-        list(APPEND libraries_path ${X})
-        continue()
-    else()
-    list(APPEND libraries_path $<TARGET_FILE:${X}>)
+        message(STATUS ${X})
+        list(APPEND LIBRARIES_PATH ${X})
     endif()
 endforeach()
 
-# get path by ${X_LIBRARY}
-list(APPEND libraries_path ${zetasql_LIBRARY} ${zookeeper_mt_LIBRARY} ${re2_LIBRARY} ${dl_LIBRARY} ${pthread_LIBRARY} ${rt_LIBRARY} ${m_LIBRARY})
+# Some third party libraries, get path by ${X_LIBRARY}
+set(THIRD_PARTY_LIBS zetasql zookeeper_mt re2 pthread rt m dl)
+foreach(X  IN LISTS THIRD_PARTY_LIBS)
+    message(STATUS ${${X}_LIBRARY})
+    find_library(${X}_LIBRARY ${X})
+    list(APPEND LIBRARIES_PATH ${${X}_LIBRARY})
+endforeach()
 
 # lib_name is the name of a library
 # all_libs save all the libraries name of absl and LLVM which find recursively
-function(get_link_libraries lib_name all_libs)
+function(get_link_libraries_by_target lib_name all_libs)
     # Determine whether it is an absl... library or a LLVM... library according to the first four digits of the ${lib_name}
     string(SUBSTRING ${lib_name} 0 4 filter)
     # Get ${lib_name}'s link library
@@ -253,28 +243,29 @@ function(get_link_libraries lib_name all_libs)
         list(FIND ${all_libs} ${Y} _index)
         # If there is no ${Y} in ${all_libs}, we recursively look for B
         if(_index EQUAL -1)
-            get_link_libraries(${Y} ${all_libs})
+            get_link_libraries_by_target(${Y} ${all_libs})
         endif()
     endforeach()
     set(${all_libs} "${${all_libs}};${lib_name}" PARENT_SCOPE)
-endfunction(get_link_libraries)
+endfunction(get_link_libraries_by_target)
 
 # ${ABSL_LIBS} and ${LLVM_LIBS} need to recursively look up
 # We need absl::statusor, but it's not mentioned
-list(APPEND ABSL_AND_LLVM ${ABSL_LIBS} absl::statusor ${LLVM_LIBS})
+list(APPEND ABSL_AND_LLVM ${ABSL_LIBS} ${LLVM_LIBS})
 foreach(X IN LISTS ABSL_AND_LLVM)
-    get_link_libraries(${X} absl_llvm_libs)
+    get_link_libraries_by_target(${X} ABSL_LLVM_LIBS)
 endforeach()
-# Remove null values
-list(REMOVE_ITEM absl_llvm_libs "")
+# Because the value of ABSL_LLVM_LIBS is changed by "ABSL_LLVM_LIBS=ABSL_LLVM_LIBS+lib_name" in recursion, and ABSL_LLVM_LIBS starts as null.
+# So there is a null value at the beginning of the final A. Remove null values here.
+list(REMOVE_ITEM ABSL_LLVM_LIBS "")
 
-# Find path of libraries in ${absl_llvm_libs}
-foreach(X  IN LISTS absl_llvm_libs)
+# Find path of libraries in ${ABSL_LLVM_LIBS}
+foreach(X  IN LISTS ABSL_LLVM_LIBS)
     get_property(_loc TARGET ${X} PROPERTY LOCATION)
-    list(APPEND absl_llvm_path ${_loc})
+    list(APPEND ABSL_LLVM_PATH ${_loc})
 endforeach()
 
-list(APPEND to_openmldbsdk $<TARGET_FILE:openmldb_api> ${absl_llvm_path} ${libraries_path})
+list(APPEND to_openmldbsdk $<TARGET_FILE:openmldb_api> ${ABSL_LLVM_PATH} ${LIBRARIES_PATH})
 
 add_custom_command(OUTPUT libopenmldbsdk.a
 COMMAND ar crsT libopenmldbsdk.a
@@ -289,15 +280,15 @@ IMPORTED_LOCATION ${CMAKE_CURRENT_BINARY_DIR}/libopenmldbsdk.a
 
 if(TESTING_ENABLE)
     add_executable(openmldb_api_test openmldb_api_test.cc)
-    target_link_libraries(openmldb_api_test openmldbsdk base_test ${BIN_LIBS} ${GTEST_LIBRARIES})
+    target_link_libraries(openmldb_api_test openmldbsdk ${BIN_LIBS} ${GTEST_LIBRARIES})
 endif()
 
 FILE(GLOB USER_HEADER 
     "${PROJECT_SOURCE_DIR}/src/sdk/openmldb_api.h"
 )
 FILE(GLOB USER_HEADER_SDK
-    "${PROJECT_SOURCE_DIR}/src/sdk/result_set.h"
-    "${PROJECT_SOURCE_DIR}/src/sdk/base_schema.h"
+    "${PROJECT_SOURCE_DIR}/hybridse/include/sdk/result_set.h"
+    "${PROJECT_SOURCE_DIR}/hybridse/include/sdk/base_schema.h"
 )
 FILE(GLOB USER_LIB
     "${PROJECT_SOURCE_DIR}/build/src/sdk/libopenmldbsdk.a"
