+++ src/sdk/openmldb_api_test.cc
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2021 4Paradigm
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 #include "openmldb_api.h"
 
 #include <unistd.h>
@@ -451,11 +467,11 @@ int main(int argc, char** argv) {
     ::openmldb::sdk::mc_ = &mc;
     FLAGS_enable_distsql = true;
     int ok = ::openmldb::sdk::mc_->SetUp(3);
+    sleep(5);
     ::openmldb::sdk::handler =
         new OpenmldbHandler(::openmldb::sdk::mc_->GetZkCluster(), ::openmldb::sdk::mc_->GetZkPath());
-    sleep(5);
     ::testing::InitGoogleTest(&argc, argv);
-    srand(time(NULL));
+    srand(time(nullptr));
     ::google::ParseCommandLineFlags(&argc, &argv, true);
     ok = RUN_ALL_TESTS();
     ::openmldb::sdk::mc_->Close();
